+++ src/main/java/DukeException.java
@@ -1,0 +1,5 @@
+class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
